validaciones de eliminar

+++ src/Vista/InterfazEliminarProducto.java
@@ -5,10 +5,16 @@
 package Vista;
 
 import Conexion.ConectarBD;
+import Controlador.ListarProducto;
 import static Vista.InterfazAdministrarProducto.bProdEditar;
 import static Vista.InterfazAdministrarProducto.bProdEliminar;
 import static Vista.InterfazAdministrarProducto.bProdRegistrar;
+import static Vista.InterfazAdministrarProducto.codProdutoLista;
 import static Vista.InterfazAdministrarProducto.codigoNumero;
+import static Vista.InterfazAdministrarProducto.jLabelMarca;
+import static Vista.InterfazAdministrarProducto.jLabelModelo;
+import static Vista.InterfazAdministrarProducto.jPanelAdministrarProducto;
+import static Vista.InterfazAdministrarProducto.offset;
 import java.awt.Image;
 import java.awt.image.BufferedImage;
 import java.sql.Statement;
@@ -34,7 +40,7 @@ public class InterfazEliminarProducto extends javax.swing.JFrame {
     static Connection conexion=null;
     static Statement sentencia=null;
     ConectarBD con = new ConectarBD();
-    
+    ListarProducto listaImagenes=new ListarProducto();
     
     /**
      * Creates new form InterfazEliminarProducto
@@ -225,18 +231,23 @@ public class InterfazEliminarProducto extends javax.swing.JFrame {
     }//GEN-LAST:event_jButtonCancelarActionPerformed
 
     private void jButtonEliminarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButtonEliminarActionPerformed
-        try {
-            sentencia=conexion.createStatement();
+        int respuesta = JOptionPane.showConfirmDialog(this, "¿Está seguro de eliminar el producto?", "Confirmación de eliminación", JOptionPane.YES_NO_OPTION);
+        if (respuesta == JOptionPane.YES_OPTION) {
+            try {
+                sentencia=conexion.createStatement();
                    
-            String sentenciaSQL1 = new String();
-            sentenciaSQL1="DELETE FROM PRODUCTO WHERE codProducto="+codigoNumero;
+                String sentenciaSQL1 = new String();
+                sentenciaSQL1="DELETE FROM PRODUCTO WHERE codProducto="+codigoNumero;
                  
-            sentencia.execute(sentenciaSQL1); 
-            JOptionPane.showMessageDialog(this, "Eliminado con exito"); 
-            dispose();
+                sentencia.execute(sentenciaSQL1); 
+                JOptionPane.showMessageDialog(this, "Eliminado con exito"); 
+                dispose();
 
-        } catch (SQLException e) {
-            Logger.getLogger(InterfazRegistrarUsuario.class.getName()).log(Level.SEVERE,null,e);
+            } catch (SQLException e) {
+                Logger.getLogger(InterfazRegistrarUsuario.class.getName()).log(Level.SEVERE,null,e);
+            }
+            offset=0;
+            listaImagenes.mostrarFotoInicio(jLabelFoto, jLabelMarca, jLabelModelo, jPanelAdministrarProducto, codProdutoLista);
         }
     }//GEN-LAST:event_jButtonEliminarActionPerformed
 public void habilitarBoton(){
